PSR 12 done - refactoring codestyle done

+++ Project/src/Dtos/Requests/CreateFoodRequest.php
@@ -6,10 +6,10 @@ class CreateFoodRequest
 {
     public function __construct(
         public string $name,
-        public int $calories,
-        public int $proteins,
-        public int $fats,
-        public int $carbs,
+        public float $calories,
+        public float $proteins,
+        public float $fats,
+        public float $carbs,
         public ?int $createdBy = null,
     ) {
     }
