+++ Project/src/Dtos/Requests/CreateMealRequest.php
@@ -10,7 +10,7 @@ class CreateMealRequest
     public function __construct(
         public int $userId,
         public int $foodId,
-        public int $amountGrams,
+        public float $amountGrams,
         public MealType $mealType,
         public DateTimeImmutable $consumedAt
     ) {
