+++ Project/src/Dtos/Responses/FoodResponse.php
@@ -7,10 +7,10 @@ class FoodResponse
     public function __construct(
         public int $id,
         public string $name,
-        public int $calories,
-        public int $proteins,
-        public int $fats,
-        public int $carbs,
+        public float $calories,
+        public float $proteins,
+        public float $fats,
+        public float $carbs,
         public ?int $createdBy = null,
     ) {
     }
